requirements.txt
langchain==1.0.5
langchain-openai==1.0.2
pymysql==1.1.0
python-dotenv==1.0.1
tiktoken==0.7.0
openai==1.3.8
